fuck

+++ src/main/scala/io/univalence/microservice/common/AlertPersonne.scala
@@ -2,7 +2,7 @@ package io.univalence.microservice.common
 
 import com.google.gson.Gson
 
-case class AlertPersonne(user_id: String,user_name: String, timestamp: Long, reason: String,coordinates : List[Float])
+case class AlertPersonne(user_id: String,user_name: String, timestamp: Long, reason: String,coordinates : List[Double])
 object AlertPersonneJson {
   val gson = new Gson()
